Move the outbound transfer logic into it's own component
