+++ client/dashboard/domains/domain-transfer/inbound-transfer.tsx
@@ -1,0 +1,5 @@
+import type { Domain } from '@automattic/api-core';
+
+export default function InboundTransfer( { domain }: { domain: Domain } ) {
+	return 'TBD';
+}
